moved components into components directory
