+++ src/components/FooterBar.js
